pyproject.toml
[project]
name = "r2framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.122.0",
    "psycopg2>=2.9.10",
    "sqlmodel>=0.0.24",
    "strands-agents[openai,otel]>=1.18.0",
    "strands-agents-tools>=0.2.16",
    "pillow>=11.2.1",
    "numpy>=2.2.6",
    "scikit-image>=0.25.2",
    "opencv-python>=4.12.0.88",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "alembic>=1.17.2",
    "aioboto3>=7.0.0",
    "scalar-fastapi>=1.4.3",
    "pytest-cov>=7.0.0",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "basedpyright>=1.37.2",
    "ruff>=0.14.14",
]
temp = [
    "opencv-python>=4.12.0.88",
    "scikit-image>=0.25.2",
]

[tool.bandit]
exclude_dirs = ["tests", "migrations"]

[tool.basedpyright]
exclude = ["**/migrations", "**/.*", "**/__pycache__", "**/tests/unit/shared"]
reportMissingTypeStubs = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportAny = false
reportExplicitAny = false
reportUnannotatedClassAttribute = false
reportImportCycles = "warning"
